Added MockDataSource and tests for post model
The MockDataSource is a way to simulate what happens with a typical service based datasource.  In particular it assigns new ids of the objects after they are created.  This is typically done by the server so the object's id matchs a database id or serverside data storage id of some kind.